File restructure, code clean-up, documentation
