+++ PrimeView.sln
@@ -12,6 +12,7 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{C370EA12-D676-4F60-8E13-D86837301F84}"
 	ProjectSection(SolutionItems) = preProject
 		src\.editorconfig = src\.editorconfig
+		README.md = README.md
 	EndProjectSection
 EndProject
 Global
