+++ src/Frontend/Filters/FilterExtensions.cs
@@ -3,7 +3,7 @@ using PrimeView.Frontend.Pages;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace PrimeView.Frontend.Tools
+namespace PrimeView.Frontend.Filters
 {
 	public static class FilterExtensions
 	{
@@ -14,10 +14,10 @@ namespace PrimeView.Frontend.Tools
 			if (filterImplementations.Count > 0)
 				source = source.Where(r => filterImplementations.Contains(r.Implementation));
 
-			return source.Where(r => 
-				r.IsMultiThreaded switch 
-				{ 
-					true => page.FilterParallelMultithreaded, 
+			return source.Where(r =>
+				r.IsMultiThreaded switch
+				{
+					true => page.FilterParallelMultithreaded,
 					_ => page.FilterParallelSinglethreaded
 				}
 				&& r.Algorithm switch
