+++ src/Frontend/Parameters/PropertyParameterMap.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq.Expressions;
 using System.Reflection;
 
-namespace PrimeView.Frontend.Tools
+namespace PrimeView.Frontend.Parameters
 {
 	public static class PropertyParameterMap
 	{
@@ -47,7 +47,7 @@ namespace PrimeView.Frontend.Tools
 		{
 			if (propertyName == null)
 				return null;
-			
+
 			string name = $"{typeof(T).Name}.{propertyName}";
 
 			return map.ContainsKey(name) ? map[name] : null;
