+++ src/Frontend/Shared/ValueTableRow.razor.css
@@ -2,8 +2,7 @@
 	width: auto;
 }
 
-th
-{
+th {
 	white-space: nowrap;
 	width: 1%;
 }
