--- src/Frontend/Tools/QueryStringParameterAttribute.cs
@@ -1,19 +1,0 @@
-﻿using System;
-
-namespace PrimeView.Frontend.Tools
-{
-	[AttributeUsage(AttributeTargets.Property, AllowMultiple = false, Inherited = true)]
-  public sealed class QueryStringParameterAttribute : Attribute
-  {
-    public QueryStringParameterAttribute()
-    {
-    }
-
-    public QueryStringParameterAttribute(string name)
-    {
-      Name = name;
-    }
-
-    public string Name { get; }
-  }
-}
