+++ src/StaticJsonReader/README.md
@@ -1,0 +1,4 @@
+﻿# Implementation notes
+
+* This project provides a "static" implementation of the [`IReportReader`](../Entities/IReportReader.cs) interface. It loads benchmark reports in JSON format from the `data` directory on the webserver. It starts with loading `report1.json`, then loads `report2.json`, and so on, until it receives an HTTP error on the request for a `report<number>.json` file.
+* The [`ExtensionMethods`](ExtensionMethods.cs) class includes a `GetStableHashCode` methodes that provides what it says on the tin. This is used to make sure that hash codes of [`Report`](../Entities/Report.cs) JSON blobs loaded from the `report<number>.json` files remain the same across different builds of the PrimeView solution('s projects). This greatly simplifies testing, as a refresh
